+++ R/CFAxis.R
@@ -488,14 +488,17 @@ CFAxis <- R6::R6Class("CFAxis",
 #'
 #' This method returns the lengths of the axes of a variable or axis.
 #'
-#' @param x The `CFVariable` or a descendant of `CFAxis`.
-#' @return Vector of axis lengths.
+#' @param x A `CFVariable` instance or a descendant of `CFAxis`.
+#' @return For a `CFVariable` instance in argument `x`, a named vector of axis
+#'   lengths, excluding any scalar axes. For a `CFAxis` descendant instance in
+#'   argument `x`, the length of the axis.
 #' @export
 #' @examples
 #' fn <- system.file("extdata", "ERA5land_Rwanda_20160101.nc", package = "ncdfCF")
 #' ds <- open_ncdf(fn)
 #' t2m <- ds[["t2m"]]
 #' dim(t2m)
+#' dim(t2m$axes[["time"]])
 dim.CFAxis <- function(x) {
   x$length
 }